Simplified button logic and offline mode
Buttons are created by a single method now, rather than having the same creation logic in each view.

Transactions are now saved to disk if they cannot be sent to the controller. When a transaction saves successfully, the transactions on disk are also uploaded.

Because of this, the way the transaction number is gotten is no-longer 100% accurate. Current transaction number is now stored on disk for this reason.